refactoring for animl-r
